streamlit/requirements.txt
google.generativeai
spacy
nltk
scikit-learn
https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0.tar.gz
sentence_transformers